REST movie API
Adding:
"index.js" file
".gitignore" file
"package-lock.json" file
"package.json" file

Adding public folder with “documentation.html” file.
 For each endpoint,  include:

A high-level description of what kind of data the endpoint returns;
The endpoint URL;
The method type (GET, POST, PUT, or DELETE);
The query parameters (parameters in the URL);
The format of data expected in the body of the request (and an example);
The format of the response data.